r125: * schroedinger/schroarith.c: * schroedinger/schroarith.h: Fix up flushing.  Expand nextbits to 32 bits, although it doesn't seem to make a performance difference.

+++ schroedinger/schroedinger/schroarith.h
@@ -105,7 +105,7 @@ struct _SchroArith {
 
   SchroBuffer *buffer;
   int offset;
-  uint8_t nextcode;
+  uint32_t nextcode;
   int nextbits;
 
   uint16_t *division_factor;
